+++ odbc-api/src/buffers.rs
@@ -6,13 +6,15 @@ mod description;
 mod indicator;
 mod text_column;
 mod text_row_set;
+mod item;
 
 pub use self::{
     bin_column::{BinColumnIt, BinColumnWriter},
-    column_with_indicator::{NullableSlice, OptWriter},
+    column_with_indicator::{NullableSlice, NullableSliceMut},
     columnar::{AnyColumnView, AnyColumnViewMut, ColumnarRowSet},
     description::{BufferDescription, BufferKind},
     indicator::Indicator,
     text_column::{CharColumn, TextColumn, TextColumnIt, TextColumnWriter, WCharColumn},
     text_row_set::TextRowSet,
+    item::Item,
 };